useAgent hook
