+++ pages/appbuilder/AppBuilderPage.tsx
@@ -1,14 +1,17 @@
 import AppBuilderContainerComponent from "@AppBuilderShared/components/shapediver/appbuilder/AppBuilderContainerComponent";
 import AppBuilderFallbackContainerComponent from "@AppBuilderShared/components/shapediver/appbuilder/AppBuilderFallbackContainerComponent";
 import MarkdownWidgetComponent from "@AppBuilderShared/components/shapediver/ui/MarkdownWidgetComponent";
-import { ComponentContext } from "@AppBuilderShared/context/ComponentContext";
+import {ComponentContext} from "@AppBuilderShared/context/ComponentContext";
 import useAppBuilderSettings from "@AppBuilderShared/hooks/shapediver/appbuilder/useAppBuilderSettings";
-import { SessionWithAppBuilderHandler, useSessionWithAppBuilder } from "@AppBuilderShared/hooks/shapediver/appbuilder/useSessionWithAppBuilder";
-import { useParameterHistory } from "@AppBuilderShared/hooks/shapediver/parameters/useParameterHistory";
-import { useSessionPropsExport } from "@AppBuilderShared/hooks/shapediver/parameters/useSessionPropsExport";
-import { useSessionPropsParameter } from "@AppBuilderShared/hooks/shapediver/parameters/useSessionPropsParameter";
+import {
+	SessionWithAppBuilderHandler,
+	useSessionWithAppBuilder,
+} from "@AppBuilderShared/hooks/shapediver/appbuilder/useSessionWithAppBuilder";
+import {useParameterHistory} from "@AppBuilderShared/hooks/shapediver/parameters/useParameterHistory";
+import {useSessionPropsExport} from "@AppBuilderShared/hooks/shapediver/parameters/useSessionPropsExport";
+import {useSessionPropsParameter} from "@AppBuilderShared/hooks/shapediver/parameters/useSessionPropsParameter";
 import useDefaultSessionDto from "@AppBuilderShared/hooks/shapediver/useDefaultSessionDto";
-import { useKeyBindings } from "@AppBuilderShared/hooks/shapediver/useKeyBindings";
+import {useKeyBindings} from "@AppBuilderShared/hooks/shapediver/useKeyBindings";
 import {
 	IAppBuilderTemplatePageContainerHints,
 	IAppBuilderTemplatePageProps,
@@ -18,8 +21,8 @@ import {
 	IAppBuilderContainer,
 	IAppBuilder,
 } from "@AppBuilderShared/types/shapediver/appbuilder";
-import { shouldUsePlatform } from "@AppBuilderShared/utils/platform/environment";
-import React, { useContext } from "react";
+import {shouldUsePlatform} from "@AppBuilderShared/utils/platform/environment";
+import React, {useContext} from "react";
 import AlertPage from "@AppBuilderShared/pages/misc/AlertPage";
 import LoaderPage from "@AppBuilderShared/pages/misc/LoaderPage";
 import AppBuilderTemplateSelector from "@AppBuilderShared/pages/templates/AppBuilderTemplateSelector";
@@ -167,14 +170,14 @@ const createContainerHints = (
  */
 export default function AppBuilderPage(props: Partial<Props>) {
 	// get default session dto, if any
-	const { defaultSessionDto } = useDefaultSessionDto(props);
+	const {defaultSessionDto} = useDefaultSessionDto(props);
 
 	// get the component context to get the correct viewport
 	const componentContext = useContext(ComponentContext);
 	const {
-		viewportComponent: { component: ViewportComponent } = {},
-		viewportOverlayWrapper: { component: ViewportOverlayWrapper } = {},
-		viewportIcons: { component: ViewportIcons } = {},
+		viewportComponent: {component: ViewportComponent} = {},
+		viewportOverlayWrapper: {component: ViewportOverlayWrapper} = {},
+		viewportIcons: {component: ViewportIcons} = {},
 	} = componentContext;
 
 	// get settings for app builder from query string
@@ -209,7 +212,7 @@ export default function AppBuilderPage(props: Partial<Props>) {
 					key={session.id}
 					props={session}
 					appBuilderOverride={settings?.appBuilderOverride}
-				/>
+				/>,
 			);
 		});
 	}
@@ -285,10 +288,10 @@ export default function AppBuilderPage(props: Partial<Props>) {
 	const show = !!sessionApi;
 
 	// use parameter history
-	useParameterHistory({ loaded: show });
+	useParameterHistory({loaded: show});
 
 	// key bindings
-	useKeyBindings({ namespace });
+	useKeyBindings({namespace});
 
 	const showMarkdown =
 		!(settings && hasSession) && // no settings or no session
@@ -322,7 +325,11 @@ export default function AppBuilderPage(props: Partial<Props>) {
 				bottom={containers.bottom}
 			>
 				{ViewportComponent && (
-					<ViewportComponent visibilitySessionIds={settings?.sessions.map((s) => s.id)} >
+					<ViewportComponent
+						visibilitySessionIds={settings?.sessions.map(
+							(s) => s.id,
+						)}
+					>
 						{ViewportOverlayWrapper && (
 							<ViewportOverlayWrapper>
 								{ViewportIcons && <ViewportIcons />}
